app/app_document_analysis/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20
httpx==0.26.0
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.3
lxml==5.1.0
